# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kubernetes FileBrowser",
        "title": "Kube-FileBrowser Swagger",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/kubeapiproxy/download": {
            "get": {
                "description": "从容器下载到本地",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes Api Proxy"
                ],
                "summary": "Download",
                "parameters": [
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "dest_paths",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "style",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "dest_paths",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "style",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "dest_paths",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "style",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "dest_paths",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "style",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "dest_paths",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "style",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/kubeapiproxy/exec": {
            "get": {
                "description": "在pod的容器中执行",
                "tags": [
                    "Kubernetes Api Proxy"
                ],
                "summary": "Exec",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "commands",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "stderr",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "stdout",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "tty",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "commands",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "stderr",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "stdout",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "tty",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "commands",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "stderr",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "stdout",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "tty",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "commands",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "stderr",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "stdout",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "tty",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "commands",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "stderr",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "stdout",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "tty",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "commands",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "stderr",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "stdout",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "tty",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "commands",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "stderr",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "stdout",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "tty",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "消息"
                }
            }
        }
    }
}